Moved code from actionButtonHandlers.js into appManager.js
Added a step to attach the add block button event listener as soon as it's rendered so it's always clickable, stoping other actions which called renderBlocks to cause the add block button to stop working
